Move types for landing page config to ui-core
