+++ ui-core/src/index.ts
@@ -1,3 +1,5 @@
 export { InvestigatorTermsOfUse } from './terms/InvestigatorTermsOfUse'
 export { ParticipantTermsOfUse } from './terms/ParticipantTermsOfUse'
 export { PrivacyPolicy } from './terms/PrivacyPolicy'
+
+export * from './types/landingPageConfig'
